import hugegraph-rpc component (#1397)
