--- hugegraph-api/src/main/java/com/baidu/hugegraph/rpc/RpcException.java
@@ -1,43 +1,0 @@
-/*
- * Copyright 2017 HugeGraph Authors
- *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with this
- * work for additional information regarding copyright ownership. The ASF
- * licenses this file to You under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-
-package com.baidu.hugegraph.rpc;
-
-import com.baidu.hugegraph.HugeException;
-
-public class RpcException extends HugeException {
-
-    private static final long serialVersionUID = -6067652498161184537L;
-
-    public RpcException(String message) {
-        super(message);
-    }
-
-    public RpcException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public RpcException(String message, Object... args) {
-        super(message, args);
-    }
-
-    public RpcException(String message, Throwable cause, Object... args) {
-        super(message, cause, args);
-    }
-}
