+++ hugegraph-dist/src/assembly/static/conf/rest-server.properties
@@ -17,10 +17,10 @@ server.role=master
 
 rpc.server_host=127.0.0.1
 rpc.server_port=8090
-rpc.server_timeout=30
+#rpc.server_timeout=30
 rpc.remote_url=127.0.0.1:8090
-rpc.client_connect_timeout=20
-rpc.client_reconnect_period=10
-rpc.client_read_timeout=40
-rpc.client_retries=3
-rpc.client_load_balancer=consistentHash
+#rpc.client_connect_timeout=20
+#rpc.client_reconnect_period=10
+#rpc.client_read_timeout=40
+#rpc.client_retries=3
+#rpc.client_load_balancer=consistentHash
